+++ src/components/SubmissionWizard/WizardHooks/WizardMapObjectsToStepsHook.tsx
@@ -1,7 +1,12 @@
-import { DisplayObjectTypes, ObjectTypes } from "constants/wizardObject"
-import { Schema, StepObject, SubmissionFolder } from "types"
+import { startCase } from "lodash"
 
-const mapObjectsToStepsHook = (submission: SubmissionFolder, objectTypesArray: Schema[]): StepObject[] => {
+import { Schema, SubmissionFolder, Workflow, WorkflowStep, MappedSteps, WorkflowSchema } from "types"
+
+const mapObjectsToStepsHook = (
+  submission: SubmissionFolder,
+  objectTypesArray: Schema[],
+  currentWorkflow: Workflow | Record<string, unknown>
+): { mappedSteps: MappedSteps[] } => {
   // Group objects by schema and status of the object
   // Sort newest first by reversing array order
   const groupedObjects = objectTypesArray
@@ -27,11 +32,10 @@ const mapObjectsToStepsHook = (submission: SubmissionFolder, objectTypesArray: S
     }, {})
 
   // Test if object type has ready or draft objects
-  const allStepItemsReady = (objectTypes: string[]) => {
+  const checkSchemaReady = (objectTypes: string[]) => {
     const foundObjects: string[] = []
-
     objectTypes.forEach(type => {
-      if (groupedObjects[type]?.drafts.length || groupedObjects[type]?.ready.length) {
+      if (groupedObjects[type]?.drafts?.length || groupedObjects[type]?.ready?.length) {
         foundObjects.push(type)
       }
     })
@@ -39,6 +43,39 @@ const mapObjectsToStepsHook = (submission: SubmissionFolder, objectTypesArray: S
     return foundObjects.length === objectTypes.length
   }
 
+  const workflowSteps: WorkflowStep[] = currentWorkflow?.steps as WorkflowStep[]
+
+  const schemaSteps =
+    workflowSteps?.length > 0
+      ? workflowSteps.map((step: WorkflowStep, index: number) => {
+          /*
+           * Filter previous steps and check whether their required schemas
+            have been saved/submitted before enabling next step
+          */
+          const previousSteps = workflowSteps.filter((step, ind) => ind < index)
+
+          const requiredSchemas = previousSteps
+            .flatMap(step => step.schemas)
+            .reduce((result: string[], schema: WorkflowSchema) => {
+              if (schema.required && schema.name !== "file") {
+                result.push(schema.name)
+              }
+              return result
+            }, [])
+
+          return {
+            ...step,
+            ["schemas"]: step.schemas.map(schema => ({
+              ...schema,
+              name: startCase(schema.name),
+              objectType: schema.name,
+              objects: groupedObjects[schema.name],
+            })),
+            actionButtonText: "Add",
+            disabled: index > 0 && !checkSchemaReady(requiredSchemas),
+          }
+        })
+      : []
   /*
    * List of accordion steps and configurations.
    * Steps are disabled by checking if previous step has been filled.
@@ -46,103 +83,22 @@ const mapObjectsToStepsHook = (submission: SubmissionFolder, objectTypesArray: S
    */
   const mappedSteps = [
     {
-      label: "Submission details",
-      stepItems: [
+      title: "Submission details",
+      schemas: [
         {
           objectType: "submissionDetails",
-          label: "Name your submission",
+          name: "Name your submission",
           objects: {
             ready: submission.submissionId ? [{ id: submission.submissionId, displayTitle: submission.name }] : [],
           },
+          required: true,
         },
       ],
       actionButtonText: "Edit",
     },
-    {
-      label: "Study, DAC and Policy",
-      stepItems: [
-        {
-          objectType: ObjectTypes.study,
-          label: DisplayObjectTypes[ObjectTypes.study],
-          objects: groupedObjects[ObjectTypes.study],
-        },
-        {
-          objectType: ObjectTypes.dac,
-          label: DisplayObjectTypes[ObjectTypes.dac],
-          objects: groupedObjects[ObjectTypes.dac],
-        },
-        {
-          objectType: ObjectTypes.policy,
-          label: DisplayObjectTypes[ObjectTypes.policy],
-          objects: groupedObjects[ObjectTypes.policy],
-        },
-      ],
-      actionButtonText: "Add",
-      disabled: submission.submissionId === "",
-    },
-    {
-      label: "Datafolder",
-      stepItems: [
-        {
-          objectType: "datafolder",
-          label: "Datafolder",
-        },
-      ],
-      actionButtonText: "Link datafolder",
-      disabled: !allStepItemsReady([ObjectTypes.study, ObjectTypes.dac, ObjectTypes.policy]), // Placeholder rule until feature is ready
-    },
-    {
-      label: "Describe",
-      stepItems: [
-        {
-          objectType: ObjectTypes.sample,
-          label: DisplayObjectTypes[ObjectTypes.sample],
-          objects: groupedObjects[ObjectTypes.sample],
-        },
-        {
-          objectType: ObjectTypes.experiment,
-          label: DisplayObjectTypes[ObjectTypes.experiment],
-          objects: groupedObjects[ObjectTypes.experiment],
-        },
-        {
-          objectType: ObjectTypes.run,
-          label: DisplayObjectTypes[ObjectTypes.run],
-          objects: groupedObjects[ObjectTypes.run],
-        },
-        {
-          objectType: ObjectTypes.analysis,
-          label: DisplayObjectTypes[ObjectTypes.analysis],
-          objects: groupedObjects[ObjectTypes.analysis],
-        },
-        {
-          objectType: ObjectTypes.dataset,
-          label: DisplayObjectTypes[ObjectTypes.dataset],
-          objects: groupedObjects[ObjectTypes.dataset],
-        },
-      ],
-      actionButtonText: "Add",
-      disabled: !allStepItemsReady([ObjectTypes.study, ObjectTypes.dac, ObjectTypes.policy]),
-    },
-    {
-      label: "Identifier and publish",
-      stepItems: [
-        {
-          objectType: "summary",
-          label: "Summary",
-        },
-      ],
-      actionButtonText: "View summary",
-      disabled: !allStepItemsReady([
-        ObjectTypes.sample,
-        ObjectTypes.run,
-        ObjectTypes.experiment,
-        ObjectTypes.dataset,
-        ObjectTypes.analysis,
-      ]),
-    },
-  ]
+  ].concat(schemaSteps)
 
-  return mappedSteps
+  return { mappedSteps }
 }
 
 export default mapObjectsToStepsHook