+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New summary page layout #791
 - New accordion stepper for submission #538
 - Update datacite's affiliation fields in cypress tests #770
+- Added changes for diffrent types of workflow when creating a submission #921
 
 ## [0.13.0] - 2021-04-7